pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "caid"
version = "0.1.0"
description = "Computer AI Design — Agentic AI framework for hardware design exploration, optimization, and interference checks"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "jinja2>=3.1.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0.0",
    # Geometry (install separately — see README):
    # cadquery>=2.4.0
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov",
]

[project.scripts]
caid = "caid_cli:main"

[tool.setuptools.packages.find]
where = ["."]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
